Support scheduled expenses.
We add support for adding and editing scheduled expenses. Scheduled expenses cause real expenses to be automatically created at given times (e.g. weekly on a specific day or monthly on a specific day).

This commit also includes a slight refactoring of some parts of the system to hopefully make it easier to maintain.