Add initial support for custom budget views.
There are still some open questions/tasks in making custom views as functional as the pre-defined views were before (e.g. allowing drill-down), but this commit is a start. Custom budget views are loaded from the backend and listed on the dashboard. The month or year can be changed (depending on whether the view is a monthly or yearly view).